# app/src/main/resources/api/openapi.yml
openapi: 3.1.0
info:
  title: News Recommendation Service - OpenAPI
  version: 0.0.1
servers:
  - url: https://localhost:4567/api/v1
...
paths:
...
  /user/register:
    post:
      tags:
        - user
      summary: Register
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        201:
          description: Success
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        422:
          description: Validation error
  /user:
    get:
      tags:
        - user
      summary: Get user information
      security:
        - BasicHttpAuthentication: []
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        401:
          description: Invalid authentication credentials
        404:
          description: Not found
    put:
      tags:
        - user
      summary: Change user information
      security:
        - BasicHttpAuthentication: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      responses:
        204:
          description: Success
        401:
          description: Invalid authentication credentials
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        422:
          description: Validation error
...
components:
  schemas:
...
    UserInfo:
      type: object
      properties:
        email:
          type: string
          example: "kruase@email.com"
        username:
          type: string
          example: "KruASe"
      required:
        - email
        - username
    UserRegister:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "kruase@email.com"
        password:
          type: string
          example: "password123"
        username:
          type: string
          minLength: 3
          example: "KruASe"
      required:
        - email
        - passwordHash
        - username
...
    ConflictError:
      type: object
      properties:
        detail:
          type: string
          example: "Email already exists"
      required:
        - detail
...
  securitySchemes:
    BasicHttpAuthentication:
      type: http
      scheme: basic
security:
  - BasicHttpAuthentication: []